+++ CHANGELOG.md
@@ -5,6 +5,12 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+## [0.0.2] - 2019-12-29
+### Changed
+- use go mod
+- make session as gate session
+- new package: gate, log
+
 ## [0.0.1] - 2018-06-17
 ### Changed
 - Move netpack func to session.go